requirements.txt
python-telegram-bot==21.4
transmission-rpc==7.0.11
requests==2.32.3
asyncio==3.4.3
python-dotenv==1.0.1
